refactor: refatorando nomes de schemas
