+++ src/modules/public/dtos/schemas/UpdateOcorrenciaSchema.ts
@@ -8,10 +8,6 @@ export const UpdateOcorrenciaSchema = z.object({
   tipoOcorrencia: z.string({ required_error: "O campo 'tipoOcorrencia' é obrigatório." }).max(100),
   artigo: z.string({ required_error: "O campo 'artigo' é obrigatório." }).max(50),
   resumo: z.string({ required_error: "O campo 'resumo' é obrigatório." }),
-  registradoPorId: z
-    .string({ required_error: "O campo 'registradoPorId' é obrigatório." })
-    .uuid({ message: "Cada policial deve ser um UUID válido" }),
-  endereco: EnderecoSchema,
 });
 
 export type UpdateOcorrenciaInput = z.infer<typeof UpdateOcorrenciaSchema>;
