+++ src/modules/public/usecases/policial/atualizar-posto-graduacao.ts
@@ -2,7 +2,7 @@ import { inject, injectable } from "tsyringe";
 import { IPolicialRepository } from "../../repositories/interfaces/IPolicialRepository";
 import { PolicialResponseDTO } from "../../dtos/response/PolicialResponseDTO";
 import AppError from "../../../../errors/AppError";
-import { AtualizarPostoGraduacaoInput } from "../../dtos/schemas/AtualizarPostoGraduacaoSchema";
+import { UpdatePostoGraduacaoInput } from "../../dtos/schemas/AtualizarPostoGraduacaoSchema";
 
 @injectable()
 export class AtualizarPostoGraduacaolUseCase {
@@ -10,10 +10,7 @@ export class AtualizarPostoGraduacaolUseCase {
     @inject("PolicialRepository") private readonly policialRepository: IPolicialRepository
   ) {}
 
-  public async execute(
-    id: string,
-    dto: AtualizarPostoGraduacaoInput
-  ): Promise<PolicialResponseDTO> {
+  public async execute(id: string, dto: UpdatePostoGraduacaoInput): Promise<PolicialResponseDTO> {
     const exists = await this.policialRepository.findById(id);
     if (!exists) {
       throw new AppError("Policial não encontrado", 404);
